reduce duplicate
