+++ benchmark/clickbench/clickbench_client.rs
@@ -1,14 +1,10 @@
 use std::{
-    fmt::Display,
     fs::File,
     io::{BufRead, BufReader},
     path::{Path, PathBuf},
-    str::FromStr,
-    sync::Arc,
-    time::{Duration, Instant},
+    time::Instant,
 };
 
-use arrow_flight::{FlightClient, flight_service_client::FlightServiceClient, sql::Any};
 use clap::{Parser, arg, command};
 use datafusion::{
     arrow::{array::RecordBatch, util::pretty},
@@ -18,230 +14,20 @@ use datafusion::{
         basic::Compression,
         file::properties::WriterProperties,
     },
-    physical_plan::{
-        ExecutionPlan, collect, display::DisplayableExecutionPlan, metrics::MetricValue,
-    },
-    prelude::{SessionConfig, SessionContext},
+    physical_plan::{collect, display::DisplayableExecutionPlan},
 };
-use futures::StreamExt;
-use liquid_cache_benchmarks::utils::assert_batch_eq;
-use liquid_cache_client::LiquidCacheTableFactory;
-use liquid_common::{
-    ParquetMode,
-    rpc::{ExecutionMetricsResponse, LiquidCacheActions},
+use liquid_cache_benchmarks::{
+    BenchmarkMode, BenchmarkResult, IterationResult, QueryResult, utils::assert_batch_eq,
 };
 use log::{debug, info};
 use mimalloc::MiMalloc;
-use object_store::ClientConfigKey;
 use owo_colors::OwoColorize;
-use prost::Message;
 use serde::Serialize;
 use std::fs::File as StdFile;
 use sysinfo::Networks;
-use tonic::transport::Channel;
-use url::Url;
 
 #[global_allocator]
 static GLOBAL: MiMalloc = MiMalloc;
-
-#[derive(Serialize)]
-struct BenchmarkResult {
-    server_url: String,
-    file_path: PathBuf,
-    query_path: PathBuf,
-    iteration: u32,
-    bench_mode: BenchmarkMode,
-    output_path: Option<PathBuf>,
-    queries: Vec<QueryResult>,
-}
-
-#[derive(Serialize)]
-struct QueryResult {
-    id: u32,
-    query: String,
-    iteration_results: Vec<IterationResult>,
-}
-
-impl QueryResult {
-    fn new(id: u32, query: String) -> Self {
-        Self {
-            id,
-            query,
-            iteration_results: Vec::new(),
-        }
-    }
-
-    fn add(&mut self, iteration_result: IterationResult) {
-        self.iteration_results.push(iteration_result);
-    }
-}
-#[derive(Serialize)]
-struct IterationResult {
-    network_traffic: u64,
-    time_millis: u64,
-    cache_cpu_time: u64,
-    cache_memory_usage: u64,
-    starting_timestamp: Duration,
-}
-
-#[derive(Clone, Debug, Default, Copy, PartialEq, Eq, Serialize)]
-enum BenchmarkMode {
-    ParquetFileserver,
-    ParquetPushdown,
-    ArrowPushdown,
-    #[default]
-    LiquidCache,
-    LiquidEagerTranscode,
-}
-
-impl BenchmarkMode {
-    async fn setup_ctx(&self, server_url: &str, file_path: &Path) -> Result<Arc<SessionContext>> {
-        let mut session_config = SessionConfig::from_env()?;
-        let table_name = "hits";
-        let current_dir = std::env::current_dir()?.to_string_lossy().to_string();
-        let table_url =
-            Url::parse(&format!("file://{}/{}", current_dir, file_path.display())).unwrap();
-
-        let mode = match self {
-            BenchmarkMode::ParquetFileserver => {
-                let ctx = Arc::new(SessionContext::new_with_config(session_config));
-                let base_url = Url::parse(server_url).unwrap();
-
-                let object_store = object_store::http::HttpBuilder::new()
-                    .with_url(base_url.clone())
-                    .with_config(ClientConfigKey::AllowHttp, "true")
-                    .build()
-                    .unwrap();
-                ctx.register_object_store(&base_url, Arc::new(object_store));
-
-                ctx.register_parquet(
-                    "hits",
-                    format!("{}/hits.parquet", server_url),
-                    Default::default(),
-                )
-                .await?;
-                return Ok(ctx);
-            }
-            BenchmarkMode::ParquetPushdown => ParquetMode::Original,
-            BenchmarkMode::ArrowPushdown => ParquetMode::Arrow,
-            BenchmarkMode::LiquidCache => ParquetMode::Liquid,
-            BenchmarkMode::LiquidEagerTranscode => ParquetMode::LiquidEagerTranscode,
-        };
-        session_config
-            .options_mut()
-            .execution
-            .parquet
-            .pushdown_filters = true;
-        let ctx = Arc::new(SessionContext::new_with_config(session_config));
-
-        let table =
-            LiquidCacheTableFactory::open_table(server_url, table_name, table_url, mode).await?;
-        ctx.register_table(table_name, Arc::new(table))?;
-        Ok(ctx)
-    }
-
-    async fn get_execution_metrics(
-        &self,
-        server_url: &str,
-        execution_plan: &Arc<dyn ExecutionPlan>,
-    ) -> ExecutionMetricsResponse {
-        match self {
-            BenchmarkMode::ParquetFileserver => {
-                // for parquet fileserver, the memory usage is the bytes scanned.
-                // It's not easy to get the memory usage as it is cached in the kernel's page cache.
-                // So the bytes scanned is the minimum cache memory usage, actual usage is slightly higher.
-                let mut plan = execution_plan;
-                while let Some(child) = plan.children().first() {
-                    plan = child;
-                }
-                if plan.name() != "ParquetExec" {
-                    // the scan is completely pruned, so the memory usage is 0
-                    return ExecutionMetricsResponse {
-                        pushdown_eval_time: 0,
-                        cache_memory_usage: 0,
-                        liquid_cache_usage: 0,
-                    };
-                }
-                let metrics = plan
-                    .metrics()
-                    .unwrap()
-                    .aggregate_by_name()
-                    .sorted_for_display()
-                    .timestamps_removed();
-
-                let mut bytes_scanned = 0;
-
-                for metric in metrics.iter() {
-                    if let MetricValue::Count { name, count } = metric.value() {
-                        if name == "bytes_scanned" {
-                            bytes_scanned = count.value();
-                        }
-                    }
-                }
-
-                ExecutionMetricsResponse {
-                    pushdown_eval_time: 0,
-                    cache_memory_usage: bytes_scanned as u64,
-                    liquid_cache_usage: 0,
-                }
-            }
-            BenchmarkMode::ParquetPushdown
-            | BenchmarkMode::ArrowPushdown
-            | BenchmarkMode::LiquidCache
-            | BenchmarkMode::LiquidEagerTranscode => {
-                let mut flight_client = get_flight_client(server_url).await;
-                let action = LiquidCacheActions::ExecutionMetrics.into();
-                let mut result_stream = flight_client.do_action(action).await.unwrap();
-                let result = result_stream.next().await.unwrap().unwrap();
-                let any = Any::decode(&*result).unwrap();
-                any.unpack::<ExecutionMetricsResponse>().unwrap().unwrap()
-            }
-        }
-    }
-
-    async fn reset_cache(&self, server_url: &str) -> Result<()> {
-        if self == &BenchmarkMode::ParquetFileserver {
-            // File server relies on OS page cache, so we don't need to reset it
-            return Ok(());
-        }
-        let mut flight_client = get_flight_client(server_url).await;
-        let action = LiquidCacheActions::ResetCache.into();
-        let mut result_stream = flight_client.do_action(action).await.unwrap();
-        let _result = result_stream.next().await.unwrap().unwrap();
-        Ok(())
-    }
-}
-
-impl Display for BenchmarkMode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{}",
-            match self {
-                BenchmarkMode::ParquetFileserver => "parquet-fileserver",
-                BenchmarkMode::ParquetPushdown => "parquet-pushdown",
-                BenchmarkMode::LiquidCache => "liquid-cache",
-                BenchmarkMode::ArrowPushdown => "arrow-pushdown",
-                BenchmarkMode::LiquidEagerTranscode => "liquid-eager-transcode",
-            }
-        )
-    }
-}
-
-impl FromStr for BenchmarkMode {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(match s {
-            "parquet-fileserver" => BenchmarkMode::ParquetFileserver,
-            "parquet-pushdown" => BenchmarkMode::ParquetPushdown,
-            "arrow-pushdown" => BenchmarkMode::ArrowPushdown,
-            "liquid-cache" => BenchmarkMode::LiquidCache,
-            "liquid-eager-transcode" => BenchmarkMode::LiquidEagerTranscode,
-            _ => return Err(format!("Invalid benchmark mode: {}", s)),
-        })
-    }
-}
 
 fn get_query(
     query_path: impl AsRef<Path>,
@@ -355,7 +141,7 @@ fn check_result_against_answer(
     Ok(())
 }
 
-#[derive(Parser)]
+#[derive(Parser, Serialize, Clone)]
 #[command(name = "ClickBench Benchmark Client")]
 struct CliArgs {
     /// Path to the query file
@@ -403,16 +189,13 @@ pub async fn main() -> Result<()> {
 
     let queries = get_query(&args.query_path, args.query)?;
     let bench_mode = &args.bench_mode;
-    let ctx = bench_mode.setup_ctx(&args.server, &args.file).await?;
+    let ctx = bench_mode
+        .setup_clickbench_ctx(&args.server, &args.file)
+        .await?;
 
     let mut benchmark_result = BenchmarkResult {
-        server_url: args.server.clone(),
-        file_path: args.file.clone(),
-        query_path: args.query_path.clone(),
-        iteration: args.iteration,
-        output_path: args.output.clone(),
-        bench_mode: *bench_mode,
-        queries: Vec::new(),
+        args: args.clone(),
+        results: Vec::new(),
     };
 
     std::fs::create_dir_all("benchmark/data/results")?;
@@ -477,7 +260,7 @@ pub async fn main() -> Result<()> {
         if args.reset_cache {
             bench_mode.reset_cache(&args.server).await?;
         }
-        benchmark_result.queries.push(query_result);
+        benchmark_result.results.push(query_result);
     }
 
     if let Some(output_path) = &args.output {
@@ -486,11 +269,4 @@ pub async fn main() -> Result<()> {
     }
 
     Ok(())
-}
-
-async fn get_flight_client(server_url: &str) -> FlightClient {
-    let endpoint = Channel::from_shared(server_url.to_string()).unwrap();
-    let channel = endpoint.connect().await.unwrap();
-    let inner_client = FlightServiceClient::new(channel);
-    FlightClient::new_from_inner(inner_client)
 }
