+++ benchmark/src/lib.rs
@@ -1,16 +1,35 @@
+use arrow_flight::sql::Any;
+use arrow_flight::{FlightClient, flight_service_client::FlightServiceClient};
+use datafusion::{error::Result, physical_plan::ExecutionPlan};
+use datafusion::{
+    physical_plan::metrics::MetricValue,
+    prelude::{SessionConfig, SessionContext},
+};
+use futures::StreamExt;
+use liquid_cache_client::LiquidCacheTableFactory;
+use liquid_cache_server::StatsCollector;
+use liquid_common::ParquetMode;
+use liquid_common::rpc::{ExecutionMetricsResponse, LiquidCacheActions};
+use liquid_parquet::LiquidCacheRef;
+use object_store::ClientConfigKey;
+use pprof::ProfilerGuard;
+use prost::Message;
+use serde::Serialize;
+use std::time::Duration;
 use std::{
-    path::PathBuf,
+    fmt::Display,
+    path::{Path, PathBuf},
+    str::FromStr,
     sync::{
         Arc, Mutex,
         atomic::{AtomicU32, AtomicUsize},
     },
 };
-pub mod utils;
-use datafusion::physical_plan::ExecutionPlan;
-use liquid_cache_server::StatsCollector;
-use liquid_parquet::LiquidCacheRef;
-use pprof::ProfilerGuard;
+use tonic::transport::Channel;
+use url::Url;
+
 pub mod admin_server;
+pub mod utils;
 
 pub struct FlameGraphReport {
     output_dir: PathBuf,
@@ -132,4 +151,275 @@ impl StatsCollector for StatsReport {
         self.cache.write_stats(&parquet_file_path).unwrap();
         log::info!("Stats saved to {}", parquet_file_path.display());
     }
+}
+
+#[derive(Clone, Debug, Default, Copy, PartialEq, Eq, Serialize)]
+pub enum BenchmarkMode {
+    ParquetFileserver,
+    ParquetPushdown,
+    ArrowPushdown,
+    #[default]
+    LiquidCache,
+    LiquidEagerTranscode,
+}
+
+impl BenchmarkMode {
+    pub async fn setup_tpch_ctx(
+        &self,
+        server_url: &str,
+        data_dir: &Path,
+    ) -> Result<Arc<SessionContext>> {
+        let mut session_config = SessionConfig::from_env()?;
+        let current_dir = std::env::current_dir()?.to_string_lossy().to_string();
+
+        let tables = [
+            "customer", "lineitem", "nation", "orders", "part", "partsupp", "region", "supplier",
+        ];
+
+        let mode = match self {
+            BenchmarkMode::ParquetFileserver => {
+                let ctx = Arc::new(SessionContext::new_with_config(session_config));
+                let base_url = Url::parse(server_url).unwrap();
+
+                let object_store = object_store::http::HttpBuilder::new()
+                    .with_url(base_url.clone())
+                    .with_config(ClientConfigKey::AllowHttp, "true")
+                    .build()
+                    .unwrap();
+                ctx.register_object_store(&base_url, Arc::new(object_store));
+
+                for table_name in tables.iter() {
+                    let table_path = Url::parse(&format!(
+                        "file://{}/{}/{}.parquet",
+                        current_dir,
+                        data_dir.display(),
+                        table_name
+                    ))
+                    .unwrap();
+                    ctx.register_parquet(*table_name, table_path, Default::default())
+                        .await?;
+                }
+                return Ok(ctx);
+            }
+            BenchmarkMode::ParquetPushdown => ParquetMode::Original,
+            BenchmarkMode::ArrowPushdown => ParquetMode::Arrow,
+            BenchmarkMode::LiquidCache => ParquetMode::Liquid,
+            BenchmarkMode::LiquidEagerTranscode => ParquetMode::LiquidEagerTranscode,
+        };
+        session_config
+            .options_mut()
+            .execution
+            .parquet
+            .pushdown_filters = true;
+        let ctx = Arc::new(SessionContext::new_with_config(session_config));
+
+        for table_name in tables.iter() {
+            let table_url = Url::parse(&format!(
+                "file://{}/{}/{}.parquet",
+                current_dir,
+                data_dir.display(),
+                table_name
+            ))
+            .unwrap();
+            let table =
+                LiquidCacheTableFactory::open_table(server_url, table_name, table_url, mode).await;
+            ctx.register_table(*table_name, Arc::new(table.unwrap()))
+                .unwrap();
+        }
+
+        Ok(ctx)
+    }
+
+    pub async fn setup_clickbench_ctx(
+        &self,
+        server_url: &str,
+        data_url: &Path,
+    ) -> Result<Arc<SessionContext>> {
+        let mut session_config = SessionConfig::from_env()?;
+        let table_name = "hits";
+        let current_dir = std::env::current_dir()?.to_string_lossy().to_string();
+        let table_url =
+            Url::parse(&format!("file://{}/{}", current_dir, data_url.display())).unwrap();
+
+        let mode = match self {
+            BenchmarkMode::ParquetFileserver => {
+                let ctx = Arc::new(SessionContext::new_with_config(session_config));
+                let base_url = Url::parse(server_url).unwrap();
+
+                let object_store = object_store::http::HttpBuilder::new()
+                    .with_url(base_url.clone())
+                    .with_config(ClientConfigKey::AllowHttp, "true")
+                    .build()
+                    .unwrap();
+                ctx.register_object_store(&base_url, Arc::new(object_store));
+
+                ctx.register_parquet(
+                    "hits",
+                    format!("{}/hits.parquet", server_url),
+                    Default::default(),
+                )
+                .await?;
+                return Ok(ctx);
+            }
+            BenchmarkMode::ParquetPushdown => ParquetMode::Original,
+            BenchmarkMode::ArrowPushdown => ParquetMode::Arrow,
+            BenchmarkMode::LiquidCache => ParquetMode::Liquid,
+            BenchmarkMode::LiquidEagerTranscode => ParquetMode::LiquidEagerTranscode,
+        };
+        session_config
+            .options_mut()
+            .execution
+            .parquet
+            .pushdown_filters = true;
+        let ctx = Arc::new(SessionContext::new_with_config(session_config));
+
+        let table =
+            LiquidCacheTableFactory::open_table(server_url, table_name, table_url, mode).await?;
+        ctx.register_table(table_name, Arc::new(table))?;
+        Ok(ctx)
+    }
+
+    pub async fn get_execution_metrics(
+        &self,
+        server_url: &str,
+        execution_plan: &Arc<dyn ExecutionPlan>,
+    ) -> ExecutionMetricsResponse {
+        match self {
+            BenchmarkMode::ParquetFileserver => {
+                // for parquet fileserver, the memory usage is the bytes scanned.
+                // It's not easy to get the memory usage as it is cached in the kernel's page cache.
+                // So the bytes scanned is the minimum cache memory usage, actual usage is slightly higher.
+                let mut plan = execution_plan;
+                while let Some(child) = plan.children().first() {
+                    plan = child;
+                }
+                if plan.name() != "ParquetExec" {
+                    // the scan is completely pruned, so the memory usage is 0
+                    return ExecutionMetricsResponse {
+                        pushdown_eval_time: 0,
+                        cache_memory_usage: 0,
+                        liquid_cache_usage: 0,
+                    };
+                }
+                let metrics = plan
+                    .metrics()
+                    .unwrap()
+                    .aggregate_by_name()
+                    .sorted_for_display()
+                    .timestamps_removed();
+
+                let mut bytes_scanned = 0;
+
+                for metric in metrics.iter() {
+                    if let MetricValue::Count { name, count } = metric.value() {
+                        if name == "bytes_scanned" {
+                            bytes_scanned = count.value();
+                        }
+                    }
+                }
+
+                ExecutionMetricsResponse {
+                    pushdown_eval_time: 0,
+                    cache_memory_usage: bytes_scanned as u64,
+                    liquid_cache_usage: 0,
+                }
+            }
+            BenchmarkMode::ParquetPushdown
+            | BenchmarkMode::ArrowPushdown
+            | BenchmarkMode::LiquidCache
+            | BenchmarkMode::LiquidEagerTranscode => {
+                let mut flight_client = get_flight_client(server_url).await;
+                let action = LiquidCacheActions::ExecutionMetrics.into();
+                let mut result_stream = flight_client.do_action(action).await.unwrap();
+                let result = result_stream.next().await.unwrap().unwrap();
+                let any = Any::decode(&*result).unwrap();
+                any.unpack::<ExecutionMetricsResponse>().unwrap().unwrap()
+            }
+        }
+    }
+
+    pub async fn reset_cache(&self, server_url: &str) -> Result<()> {
+        if self == &BenchmarkMode::ParquetFileserver {
+            // File server relies on OS page cache, so we don't need to reset it
+            return Ok(());
+        }
+        let mut flight_client = get_flight_client(server_url).await;
+        let action = LiquidCacheActions::ResetCache.into();
+        let mut result_stream = flight_client.do_action(action).await.unwrap();
+        let _result = result_stream.next().await.unwrap().unwrap();
+        Ok(())
+    }
+}
+
+impl Display for BenchmarkMode {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                BenchmarkMode::ParquetFileserver => "parquet-fileserver",
+                BenchmarkMode::ParquetPushdown => "parquet-pushdown",
+                BenchmarkMode::LiquidCache => "liquid-cache",
+                BenchmarkMode::ArrowPushdown => "arrow-pushdown",
+                BenchmarkMode::LiquidEagerTranscode => "liquid-eager-transcode",
+            }
+        )
+    }
+}
+
+impl FromStr for BenchmarkMode {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(match s {
+            "parquet-fileserver" => BenchmarkMode::ParquetFileserver,
+            "parquet-pushdown" => BenchmarkMode::ParquetPushdown,
+            "arrow-pushdown" => BenchmarkMode::ArrowPushdown,
+            "liquid-cache" => BenchmarkMode::LiquidCache,
+            "liquid-eager-transcode" => BenchmarkMode::LiquidEagerTranscode,
+            _ => return Err(format!("Invalid benchmark mode: {}", s)),
+        })
+    }
+}
+
+async fn get_flight_client(server_url: &str) -> FlightClient {
+    let endpoint = Channel::from_shared(server_url.to_string()).unwrap();
+    let channel = endpoint.connect().await.unwrap();
+    let inner_client = FlightServiceClient::new(channel);
+    FlightClient::new_from_inner(inner_client)
+}
+
+#[derive(Serialize)]
+pub struct BenchmarkResult<T: Serialize> {
+    pub args: T,
+    pub results: Vec<QueryResult>,
+}
+
+#[derive(Serialize)]
+pub struct QueryResult {
+    id: u32,
+    query: String,
+    iteration_results: Vec<IterationResult>,
+}
+
+impl QueryResult {
+    pub fn new(id: u32, query: String) -> Self {
+        Self {
+            id,
+            query,
+            iteration_results: Vec::new(),
+        }
+    }
+
+    pub fn add(&mut self, iteration_result: IterationResult) {
+        self.iteration_results.push(iteration_result);
+    }
+}
+#[derive(Serialize)]
+pub struct IterationResult {
+    pub network_traffic: u64,
+    pub time_millis: u64,
+    pub cache_cpu_time: u64,
+    pub cache_memory_usage: u64,
+    pub starting_timestamp: Duration,
 }
